Change notification charge to show decimal places

+++ src/EA.Iws.Web/Areas/NotificationApplication/ViewModels/NotificationApplication/NotificationOverviewViewModel.cs
@@ -30,7 +30,7 @@
 
         public SubmitSideBarViewModel SubmitSideBarViewModel { get; set; }
 
-        public int NotificationCharge { get; set; }
+        public decimal NotificationCharge { get; set; }
 
         public bool CanEditNotification { get; set; }
